+++ doc/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 Version 0.7.6
 -------------
 
-* ``convert``: add ``--remove-all-metadata`` command
+* ``convert``: add ``--remove-all-metadata`` and ``--update-metadata`` commands
 
 Version 0.7.5 (2016-09-01)
 --------------------------