+++ doc/changelog.rst
@@ -5,6 +5,7 @@ Version 0.7.6
 -------------
 
 * ``convert``: add ``--remove-all-metadata`` and ``--update-metadata`` commands
+* Add ``unit`` metadata: ``unit=byte`` formats samples as file sizes
 
 Version 0.7.5 (2016-09-01)
 --------------------------